@@ -3,43 +3,46 @@
   "description": "Paragon Pattern Library Documentation",
   "version": "1.0.0",
   "dependencies": {
-    "@docsearch/react": "^3.1.0",
+    "@docsearch/react": "^3.8",
     "@edx/brand-edx.org": "^2.1.2",
-    "@openedx/brand-openedx": "^1.2.2",
     "@fortawesome/free-solid-svg-icons": "^6.1.1",
-    "@mdx-js/mdx": "^1.6.22",
-    "@mdx-js/react": "^1.6.22",
+    "@mdx-js/mdx": "^2",
+    "@mdx-js/react": "^2",
+    "@openedx/brand-openedx": "^1.2.2",
     "analytics-node": "^6.0.0",
     "axios": "^0.28.0",
     "classnames": "^2.3.1",
     "color": "^4.2.3",
-    "gatsby": "^4.23.1",
-    "gatsby-plugin-manifest": "^4.17.0",
-    "gatsby-plugin-mdx": "^3.17.0",
+    "gatsby": "^5.14",
+    "gatsby-plugin-manifest": "^5.14",
+    "gatsby-plugin-mdx": "^5.14",
+    "gatsby-plugin-mdx-source-name": "^1.0.1",
     "gatsby-plugin-react-axe": "^0.5.0",
-    "gatsby-plugin-react-helmet": "^5.17.0",
-    "gatsby-plugin-sass": "^5.17.0",
+    "gatsby-plugin-react-helmet": "^6.14",
+    "gatsby-plugin-sass": "^6.14",
     "gatsby-plugin-segment-js": "^3.7.1",
-    "gatsby-source-filesystem": "^4.17.0",
-    "gatsby-transformer-react-docgen": "^7.17.0",
+    "gatsby-source-filesystem": "^5.14",
+    "gatsby-transformer-react-docgen": "^8.14",
     "localforage": "^1.10.0",
     "lodash.debounce": "^4.0.8",
     "prism-react-renderer": "^1.3.3",
     "prop-types": "^15.8.1",
     "query-string": "^8.1.0",
-    "react": "^17.0.2",
+    "react": "^18",
     "react-docgen": "^5.4.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18",
     "react-focus-on": "^3.6.0",
     "react-helmet": "^6.1.0",
-    "react-intl": "^5.25.1",
+    "react-intl": "^6.4",
     "react-live": "^2.4.1",
+    "react-markdown": "^9.0.1",
     "rehype-autolink-headings": "^5.1.0",
+    "rehype-mdx-code-props": "^3.0.1",
     "rehype-slug": "^4.0.1",
     "sass": "^1.53.0",
     "sass-loader": "12.6.0",
-    "uuid": "^9.0.0",
-    "slugify": "^1.6.5"
+    "slugify": "^1.6.5",
+    "uuid": "^9.0.0"
   },
   "keywords": [
     "paragon",
