@@ -1,8 +1,12 @@
-require('dotenv').config({
-  path: `.env.${process.env.NODE_ENV}`,
-});
-const rehypeSlugPlugin = require('rehype-slug');
-const rehypeAutolinkHeadingsPlugin = require('rehype-autolink-headings');
+import { dirname } from 'path';
+import { fileURLToPath } from 'url';
+import dotenv from 'dotenv';
+dotenv.config({ path: `.env.${process.env.NODE_ENV}` });
+import rehypeSlugPlugin from 'rehype-slug';
+import rehypeAutolinkHeadingsPlugin from 'rehype-autolink-headings';
+import rehypeMdxCodeProps from 'rehype-mdx-code-props';
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
 
 const segmentPlugin = {
   resolve: 'gatsby-plugin-segment-js',
@@ -31,9 +35,13 @@ const plugins = [
           namedExport: false,
         },
       },
+      sassOptions: {
+        silenceDeprecations: ['abs-percent', 'color-functions', 'import', 'mixed-decls', 'global-builtin'],
+      },
     },
   },
   'gatsby-plugin-react-helmet',
+  'gatsby-plugin-mdx-source-name',
   {
     resolve: 'gatsby-plugin-manifest',
     options: {
@@ -55,6 +63,13 @@ const plugins = [
   {
     resolve: 'gatsby-source-filesystem',
     options: {
+      path: `${__dirname}/src/pages`,
+      name: 'pages',
+    },
+  },
+  {
+    resolve: 'gatsby-source-filesystem',
+    options: {
       path: `${__dirname}/../CHANGELOG.md`,
       name: 'changelog',
     },
@@ -66,40 +81,35 @@ const plugins = [
     resolve: 'gatsby-plugin-mdx',
     options: {
       extensions: ['.mdx', '.md'],
-      defaultLayouts: {
-        components: require.resolve(
-          './src/templates/component-page-template.tsx',
-        ),
-        default: require.resolve(
-          './src/templates/default-mdx-page-template.tsx',
-        ),
-      },
-      rehypePlugins: [
-        rehypeSlugPlugin,
-        [
-          rehypeAutolinkHeadingsPlugin,
-          {
-            behavior: 'append',
-            content: {
-              type: 'element',
-              tagName: 'span',
-              properties: {
-                className: 'pgn-doc__anchor',
+      mdxOptions: {
+        rehypePlugins: [
+          rehypeSlugPlugin,
+          [rehypeMdxCodeProps, { tagName: 'code' }],
+          [
+            rehypeAutolinkHeadingsPlugin,
+            {
+              behavior: 'append',
+              content: {
+                type: 'element',
+                tagName: 'span',
+                properties: {
+                  className: 'pgn-doc__anchor',
+                },
+                children: [
+                  { type: 'text', value: '#' },
+                ],
               },
-              children: [
-                { type: 'text', value: '#' },
-              ],
             },
-          },
+          ],
         ],
-      ],
+      },
     },
   },
   {
     resolve: 'gatsby-plugin-page-creator',
     options: {
       path: `${__dirname}/src/pages`,
-      ignore: ['insights.tsx'],
+      ignore: ['insights.tsx', '**/*.(md|mdx)'],
     },
   },
 ];
@@ -112,7 +122,7 @@ if (process.env && process.env.FEATURE_ENABLE_AXE) {
   plugins.push(axePlugin);
 }
 
-module.exports = {
+export default {
   siteMetadata: {
     title: 'Paragon Design System',
     description: 'Technical documentation for the Paragon Design System.',
@@ -121,4 +131,7 @@ module.exports = {
   // Match the location of the site on github pages if no path prefix is specified
   pathPrefix: process.env.PATH_PREFIX || '',
   plugins,
+  flags: {
+    FAST_DEV: true,
+  },
 };
