@@ -48,7 +48,8 @@ describe('<Breadcrumb />', () => {
     expect(screen.getAllByRole('presentation').length).toBe(2);
   });
 
-  it('fires the passed in click handler', () => {
+  it('fires the passed in click handler', async () => {
+    const user = userEvent.setup();
     const clickHandler = jest.fn();
     render(<Breadcrumb {...baseProps} clickHandler={clickHandler} />);
 
@@ -56,7 +57,7 @@ describe('<Breadcrumb />', () => {
     const links = screen.queryAllByRole('link');
     expect(listItems.length).toBe(baseProps.links.length);
 
-    userEvent.click(links[0]);
+    await user.click(links[0]);
     expect(clickHandler).toHaveBeenCalled();
   });
 
