@@ -14,21 +14,23 @@ describe('<Button />', () => {
     expect(button).toBeInTheDocument();
   });
 
-  it('puts focus on button on click', () => {
+  it('puts focus on button on click', async () => {
+    const user = userEvent.setup();
     const { getByText } = render(<Button {...defaultProps} />);
     const button = getByText(defaultProps.label);
 
     expect(button).not.toHaveFocus();
-    userEvent.click(button);
+    await user.click(button);
     expect(button).toHaveFocus();
   });
 
-  it('calls onClick prop on click', () => {
+  it('calls onClick prop on click', async () => {
+    const user = userEvent.setup();
     const onClickSpy = jest.fn();
     const { getByText } = render(<Button {...defaultProps} onClick={onClickSpy} />);
     const button = getByText(defaultProps.label);
 
-    userEvent.click(button);
+    await user.click(button);
     expect(onClickSpy).toHaveBeenCalledTimes(1);
   });
 });
