@@ -89,59 +89,64 @@ describe('render behavior', () => {
     expect(screen.getByDisplayValue('Test Value')).toBeInTheDocument();
   });
 
-  it('renders component with options', () => {
+  it('renders component with options', async () => {
+    const user = userEvent.setup();
     const { getByTestId, queryAllByTestId } = render(<FormAutosuggestTestComponent />);
     const input = getByTestId('autosuggest-textbox-input');
-    userEvent.click(input);
+    await user.click(input);
     const list = queryAllByTestId('autosuggest-optionitem');
     expect(list.length).toBe(3);
   });
 
-  it('renders with value required error msg', () => {
+  it('renders with value required error msg', async () => {
+    const user = userEvent.setup();
     const { getByText, getByTestId } = render(<FormAutosuggestTestComponent isValueRequired />);
     const input = getByTestId('autosuggest-textbox-input');
 
     // if you click into the input and click outside, you should see the error message
-    userEvent.click(input);
-    userEvent.click(document.body);
+    await user.click(input);
+    await user.click(document.body);
 
     const formControlFeedback = getByText('Example value required error message');
 
     expect(formControlFeedback).toBeInTheDocument();
   });
 
-  it('renders with selection required error msg', () => {
+  it('renders with selection required error msg', async () => {
+    const user = userEvent.setup();
     const { getByText, getByTestId } = render(<FormAutosuggestTestComponent isSelectionRequired />);
     const input = getByTestId('autosuggest-textbox-input');
 
     // if you click into the input and click outside, you should see the error message
-    userEvent.click(input);
-    userEvent.type(input, '1');
-    userEvent.click(document.body);
+    await user.click(input);
+    await user.type(input, '1');
+    await user.click(document.body);
 
     const formControlFeedback = getByText('Example selection required error message');
 
     expect(formControlFeedback).toBeInTheDocument();
   });
 
-  it('renders with custom error msg', () => {
+  it('renders with custom error msg', async () => {
+    const user = userEvent.setup();
     const { getByText, getByTestId } = render(<FormAutosuggestTestComponent hasCustomError />);
     const input = getByTestId('autosuggest-textbox-input');
 
     // if you click into the input and click outside, you should see the error message
-    userEvent.click(input);
-    userEvent.click(document.body);
+    await user.click(input);
+    await user.click(document.body);
 
     const formControlFeedback = getByText('Example custom error message');
 
     expect(formControlFeedback).toBeInTheDocument();
   });
 
-  it('renders component with options that all have IDs', () => {
+  it('renders component with options that all have IDs', async () => {
+    const user = userEvent.setup();
     const { getByTestId, getAllByTestId } = render(<FormAutosuggestTestComponent />);
     const input = getByTestId('autosuggest-textbox-input');
 
-    userEvent.click(input);
+    await user.click(input);
     const optionItemIds = getAllByTestId('autosuggest-optionitem').map(item => item.id);
 
     expect(optionItemIds).not.toContain(null);
@@ -154,12 +159,13 @@ describe('render behavior', () => {
     expect(getByTestId('autosuggest-screen-reader-options-count').getAttribute('aria-live')).toEqual('assertive');
   });
 
-  it('displays correct amount of options found to screen readers', () => {
+  it('displays correct amount of options found to screen readers', async () => {
+    const user = userEvent.setup();
     const { getByText, getByTestId } = render(<FormAutosuggestTestComponent />);
     const input = getByTestId('autosuggest-textbox-input');
 
     expect(getByText('0 options found')).toBeInTheDocument();
-    userEvent.click(input);
+    await user.click(input);
 
     expect(getByText('3 options found')).toBeInTheDocument();
   });
@@ -172,167 +178,180 @@ describe('render behavior', () => {
 });
 
 describe('controlled behavior', () => {
-  it('sets input value based on clicked option', () => {
+  it('sets input value based on clicked option', async () => {
+    const user = userEvent.setup();
     const { getByText, getByTestId } = render(<FormAutosuggestTestComponent />);
     const input = getByTestId('autosuggest-textbox-input');
 
-    userEvent.click(input);
+    await user.click(input);
     const menuItem = getByText('Option 1');
-    userEvent.click(menuItem);
+    await user.click(menuItem);
 
     expect(input.value).toEqual('Option 1');
   });
 
-  it('calls onChange based on clicked option', () => {
+  it('calls onChange based on clicked option', async () => {
+    const user = userEvent.setup();
     const onChange = jest.fn();
     const { getByText, getByTestId } = render(<FormAutosuggestTestComponent onChange={onChange} />);
     const input = getByTestId('autosuggest-textbox-input');
 
-    userEvent.click(input);
+    await user.click(input);
     const menuItem = getByText('Option 1');
-    userEvent.click(menuItem);
+    await user.click(menuItem);
 
     expect(onChange).toHaveBeenCalledWith({ selectionId: 'option-1-id', selectionValue: 'Option 1', userProvidedText: 'Option 1' });
     expect(onChange).toHaveBeenCalledTimes(1);
   });
 
-  it('calls onChange when the textbox is cleared', () => {
+  it('calls onChange when the textbox is cleared', async () => {
+    const user = userEvent.setup();
     const onChange = jest.fn();
     const { getByTestId } = render(<FormAutosuggestTestComponent onChange={onChange} />);
     const input = getByTestId('autosuggest-textbox-input');
 
-    userEvent.type(input, '1');
-    userEvent.type(input, '{backspace}');
+    await user.type(input, '1');
+    await user.type(input, '{backspace}');
 
     expect(onChange).toHaveBeenCalledWith({ selectionId: '', selectionValue: '', userProvidedText: '' });
   });
 
-  it('calls the function passed to onClick when an option with it is selected', () => {
+  it('calls the function passed to onClick when an option with it is selected', async () => {
+    const user = userEvent.setup();
     const onClick = jest.fn();
     const { getByText, getByTestId } = render(<FormAutosuggestTestComponent onClick={onClick} />);
     const input = getByTestId('autosuggest-textbox-input');
 
-    userEvent.click(input);
+    await user.click(input);
     const menuItem = getByText('Option 2');
-    userEvent.click(menuItem);
+    await user.click(menuItem);
 
     expect(onClick).toHaveBeenCalledTimes(1);
   });
 
-  it('does not call onClick when an option without it is selected', () => {
+  it('does not call onClick when an option without it is selected', async () => {
+    const user = userEvent.setup();
     const onClick = jest.fn();
     const { getByText, getByTestId } = render(<FormAutosuggestTestComponent onClick={onClick} />);
     const input = getByTestId('autosuggest-textbox-input');
 
-    userEvent.click(input);
+    await user.click(input);
     const menuItem = getByText('Option 1');
-    userEvent.click(menuItem);
+    await user.click(menuItem);
 
     expect(onClick).toHaveBeenCalledTimes(0);
   });
 
-  it('should set the correct activedescendant', () => {
+  it('should set the correct activedescendant', async () => {
+    const user = userEvent.setup();
     const { getByTestId, getAllByTestId } = render(<FormAutosuggestTestComponent />);
     const input = getByTestId('autosuggest-textbox-input');
 
-    userEvent.click(input);
+    await user.click(input);
     const expectedOptionId = getAllByTestId('autosuggest-optionitem')[0].id;
-    userEvent.keyboard('{arrowdown}');
+    await user.keyboard('{arrowdown}');
 
     expect(input.getAttribute('aria-activedescendant')).toEqual(expectedOptionId);
   });
 
-  it('filters dropdown based on typed field value with one match', () => {
+  it('filters dropdown based on typed field value with one match', async () => {
+    const user = userEvent.setup();
     const { getByTestId, queryAllByTestId } = render(<FormAutosuggestTestComponent />);
     const input = getByTestId('autosuggest-textbox-input');
 
-    userEvent.click(input);
-    userEvent.type(input, 'Option 1');
+    await user.click(input);
+    await user.type(input, 'Option 1');
 
     const list = queryAllByTestId('autosuggest-optionitem');
     expect(list.length).toBe(1);
   });
 
-  it('toggles options list', () => {
+  it('toggles options list', async () => {
+    const user = userEvent.setup();
     const { getByTestId, queryAllByTestId } = render(<FormAutosuggestTestComponent />);
     const dropdownBtn = getByTestId('autosuggest-iconbutton');
 
-    userEvent.click(dropdownBtn);
+    await user.click(dropdownBtn);
     const list = queryAllByTestId('autosuggest-optionitem');
     expect(list.length).toBe(3);
 
-    userEvent.click(dropdownBtn);
+    await user.click(dropdownBtn);
     const updatedList = queryAllByTestId('autosuggest-optionitem');
     expect(updatedList.length).toBe(0);
 
-    userEvent.click(dropdownBtn);
+    await user.click(dropdownBtn);
     const reopenedList = queryAllByTestId('autosuggest-optionitem');
     expect(reopenedList.length).toBe(3);
   });
 
-  it('filters dropdown based on typed field value with multiple matches', () => {
+  it('filters dropdown based on typed field value with multiple matches', async () => {
+    const user = userEvent.setup();
     const { getByTestId, queryAllByTestId } = render(<FormAutosuggestTestComponent />);
     const input = getByTestId('autosuggest-textbox-input');
 
-    userEvent.click(input);
-    userEvent.type(input, '1');
+    await user.click(input);
+    await user.type(input, '1');
 
     const list = queryAllByTestId('autosuggest-optionitem');
     expect(list.length).toBe(2);
   });
 
-  it('closes options list on click outside', () => {
+  it('closes options list on click outside', async () => {
+    const user = userEvent.setup();
     const { getByTestId, queryAllByTestId } = render(<FormAutosuggestTestComponent />);
     const input = getByTestId('autosuggest-textbox-input');
 
-    userEvent.click(input);
+    await user.click(input);
     const list = queryAllByTestId('autosuggest-optionitem');
     expect(list.length).toBe(3);
 
-    userEvent.click(document.body);
+    await user.click(document.body);
     const updatedList = queryAllByTestId('autosuggest-optionitem');
     expect(updatedList.length).toBe(0);
   });
 
-  it('updates screen reader option count based on typed field value with multiple matches', () => {
+  it('updates screen reader option count based on typed field value with multiple matches', async () => {
+    const user = userEvent.setup();
     const { getByText, getByTestId } = render(<FormAutosuggestTestComponent />);
     const input = getByTestId('autosuggest-textbox-input');
 
     expect(getByText('0 options found')).toBeInTheDocument();
-    userEvent.click(input);
+    await user.click(input);
 
     expect(getByText('3 options found')).toBeInTheDocument();
 
-    userEvent.click(input);
-    userEvent.type(input, '1');
+    await user.click(input);
+    await user.type(input, '1');
 
     expect(getByText('2 options found')).toBeInTheDocument();
   });
 
-  it('closes options list when tabbed out and the input is no longer active', () => {
+  it('closes options list when tabbed out and the input is no longer active', async () => {
+    const user = userEvent.setup();
     const { getByTestId, queryAllByTestId } = render(<FormAutosuggestTestComponent />);
     const input = getByTestId('autosuggest-textbox-input');
 
-    userEvent.click(input);
+    await user.click(input);
     expect(document.activeElement).toBe(getByTestId('autosuggest-textbox-input'));
 
     const list = queryAllByTestId('autosuggest-optionitem');
     expect(list.length).toBe(3);
 
-    userEvent.tab();
+    await user.tab();
     expect(document.activeElement).not.toBe(getByTestId('autosuggest-textbox-input'));
 
     const updatedList = queryAllByTestId('autosuggest-optionitem');
     expect(updatedList.length).toBe(0);
   });
 
-  it('check focus on input after esc', () => {
+  it('check focus on input after esc', async () => {
+    const user = userEvent.setup();
     const { getByTestId } = render(<FormAutosuggestTestComponent />);
     const input = getByTestId('autosuggest-textbox-input');
     const dropdownBtn = getByTestId('autosuggest-iconbutton');
-    userEvent.click(dropdownBtn);
+    await user.click(dropdownBtn);
 
-    userEvent.keyboard('{esc}');
+    await user.keyboard('{Escape}');
 
     expect(input.matches(':focus')).toBe(true);
   });
