@@ -8,7 +8,6 @@ import React, {
   useMemo,
   useRef,
 } from 'react';
-import PropTypes from 'prop-types';
 import { Link } from 'gatsby';
 import axios from 'axios';
 import classNames from 'classnames';
@@ -121,8 +120,8 @@ export interface ICodeBlock {
 
 function CodeBlock({
   children,
-  className,
-  live,
+  className = '',
+  live = false,
 }: ICodeBlock) {
   const intl = useIntl();
   const language: any = className ? className.replace(/language-/, '') : 'jsx';
@@ -135,6 +134,11 @@ function CodeBlock({
     navigator.clipboard.writeText(codeExample);
     setShowToast(true);
   };
+
+  if (className === '' && typeof children === 'string' && !children.includes('\n')) {
+    // This is an inline code block. Don't use syntax highlighting.
+    return <code>{children}</code>;
+  }
 
   if (live) {
     return (
@@ -204,16 +208,5 @@ function CodeBlock({
     </Highlight>
   );
 }
-
-CodeBlock.propTypes = {
-  children: PropTypes.string.isRequired,
-  className: PropTypes.string,
-  live: PropTypes.bool,
-};
-
-CodeBlock.defaultProps = {
-  live: false,
-  className: '',
-};
 
 export default CodeBlock;
