@@ -1,10 +1,9 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 
 export interface ILinkedHeading {
   h: string,
-  children: React.ReactNode,
-  id: string,
+  children?: React.ReactNode,
+  id?: string,
 }
 
 type HeadingTag = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
@@ -13,7 +12,7 @@ function LinkedHeading({
   h,
   children,
   id,
-}: ILinkedHeading) {
+}: ILinkedHeading & JSX.IntrinsicElements['h2']): React.ReactNode {
   const H = `h${h}` as HeadingTag;
 
   return (
@@ -22,11 +21,5 @@ function LinkedHeading({
     </H>
   );
 }
-
-LinkedHeading.propTypes = {
-  h: PropTypes.string.isRequired,
-  children: PropTypes.node.isRequired,
-  id: PropTypes.string.isRequired,
-};
 
 export default LinkedHeading;
