@@ -25,7 +25,7 @@ export default function StatusPage({ pageContext }) {
         <h3>Components Status</h3>
         <StaticQuery
           query={graphql`query ComponentStatusQuery {
-            allMdx(filter: {frontmatter: {type: {eq: "component"}}}, sort: {fields: frontmatter___title}) {
+            allMdx(filter: {frontmatter: {type: {eq: "component"}}}, sort: {frontmatter: {title: ASC}}) {
               nodes {
                 frontmatter {
                   designStatus
